+++ lib/ui/views/nft_creation/layouts/components/nft_creation_form_sheet.dart
@@ -1,10 +1,13 @@
 import 'package:aewallet/application/settings/theme.dart';
+import 'package:aewallet/ui/util/dimens.dart';
 import 'package:aewallet/ui/util/styles.dart';
 import 'package:aewallet/ui/views/nft/layouts/components/nft_header.dart';
 import 'package:aewallet/ui/views/nft_creation/bloc/provider.dart';
 import 'package:aewallet/ui/views/nft_creation/bloc/state.dart';
 import 'package:aewallet/ui/views/nft_creation/layouts/nft_creation_process_sheet.dart';
+import 'package:aewallet/ui/widgets/components/app_button_tiny.dart';
 import 'package:aewallet/ui/widgets/components/icons.dart';
+import 'package:aewallet/ui/widgets/components/popup_dialog.dart';
 import 'package:contained_tab_bar_view/contained_tab_bar_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/localizations.dart';
@@ -58,6 +61,15 @@ class _NftCreationFormSheetState extends ConsumerState<NftCreationFormSheet> {
               children: <Widget>[
                 NFTHeader(
                   currentNftCategoryIndex: nftCreation.currentNftCategoryIndex,
+                  onPressBack: () async {
+                    await showDialog(
+                      context: context,
+                      barrierDismissible: false,
+                      builder: (BuildContext context) {
+                        return const _NFTCreationBackPopup();
+                      },
+                    );
+                  },
                 ),
                 Divider(
                   height: 2,
@@ -169,6 +181,63 @@ class _NftCreationFormSheetState extends ConsumerState<NftCreationFormSheet> {
             ),
           ),
         ),
+      ),
+    );
+  }
+}
+
+class _NFTCreationBackPopup extends ConsumerWidget {
+  const _NFTCreationBackPopup();
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final localizations = AppLocalizations.of(context)!;
+    final theme = ref.watch(ThemeProviders.selectedTheme);
+
+    return PopupDialog(
+      title: Text(
+        localizations.exitNFTCreationProcessTitle,
+        style: theme.textStyleSize16W400Primary,
+      ),
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Text(
+            localizations.exitNFTCreationProcessSubtitle,
+            style: theme.textStyleSize14W200Primary,
+          ),
+          const SizedBox(
+            height: 20,
+          ),
+          Row(
+            children: [
+              AppButtonTiny(
+                AppButtonTinyType.primary,
+                localizations.no,
+                Dimens.buttonTopDimens,
+                onPressed: () {
+                  Navigator.pop(context);
+                },
+              ),
+              AppButtonTiny(
+                AppButtonTinyType.primary,
+                localizations.yes,
+                Dimens.buttonTopDimens,
+                onPressed: () {
+                  /**
+                   * Go back 2 times:
+                   * - Popup
+                   * - Nft form creation
+                   */
+                  var count = 0;
+                  Navigator.popUntil(context, (route) {
+                    return count++ == 2;
+                  });
+                },
+              ),
+            ],
+          ),
+        ],
       ),
     );
   }
